implemented question 1
